dom/workers/test/serviceworkers/test_install_event.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 94048 - test install event.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function simpleRegister() {
    var p = navigator.serviceWorker.register("worker.js", { scope: "./" });
    return p;
  }

  function nextRegister(reg) {
    var p = navigator.serviceWorker.register("install_event_worker.js", { scope: "./" });

    return new Promise(function(resolve, reject) {
      reg.onupdatefound = function(e) {
        ok(true, "Received onupdatefound");
        resolve();
      };
    });
  }

  function unregister() {
    return navigator.serviceWorker.getRegistration("./").then(function(reg) {
      return reg.unregister();
    });
  }

  function runTest() {
    simpleRegister()
      .then(nextRegister)
      .then(unregister)
      .then(function() {
        SimpleTest.finish();
      }).catch(function(e) {
        ok(false, "Some test failed with error " + e);
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
